Create Info_Port_Net_Scan.py
This Python script is a simple tool that takes an IP address as input and provides network and system-related information. It is designed to retrieve essential details about a specific IP address, including hostname, operating system details, and hardware information. The tool is useful for obtaining basic network and system insights without the need for external libraries or tools.

## Features

- Retrieves the hostname of the provided IP address.
- Displays the operating system details, including the OS name and version.
- Provides hardware information such as CPU details (if available), total RAM, and total HDD storage related to the provided IP address.
-Display Open Ports